requirements.txt
google-generativeai
streamlit
chromadb
langchain
python-dotenv
pypdf
tiktoken